Example of styles improvements
