+++ src/App.scss
@@ -1,4 +1,4 @@
-@import './shared/components/main-styles';
+@import './shared/main-styles';
 
 body {
   background-color: $bg-color;
