+++ src/index.scss
@@ -1,1 +1,4 @@
-
+body {
+  margin: 0;
+  font-family: Avenir, Helvetica, Arial, sans-serif;
+}
