--- src/screens/ToDoList/ToDoListItem/styl.scss
@@ -1,50 +1,0 @@
-@import './../../../shared/components/main-styles.scss';
-
-.row-item {
-  @include row-form;
-  margin-left: 50px;
-  align-items: center;
-}
-
-.row-item-completed {
-  @include row-form;
-  position: relative;
-  @include button-styles;
-  opacity: .5;
-  text-decoration: line-through; 
-  margin-left: 50px; 
-  align-items: center;
-}
-
-.row {
-  @include row-form;
-  align-items: center;
-}
-
-.delete-button {
-  background-color: white;
-  font-size: 15px;
-  border: none;
-  color: red;
-  margin-top: 5px;
-  padding: 0px;
-  margin-left: 14px;
-  @include button-styles;
-}
-
-.delete-button:hover {
-  background-color: white;
-  transform: scale(1.5);
-}
-
-.checkbox :hover {
-  transform: scale(1.5);
-}
-
-.checkbox {
-  color: green;
-  margin-right: 14px;
-  margin-top: 5px;
-  transition: transform 0.5s ease-in-out;
-  @include button-styles;
-}
