--- src/shared/components/Input/Input/styles.scss
@@ -1,19 +1,0 @@
-@import './../../main-styles.scss';
-
-.input-add-item {
-  border: 2px solid lightgray;
-  border-radius: 8px;
-  height: 30px;
-  color: $input-text-color;
-  width: 290px;
-  outline: 0;
-  margin-left: 10px;
-  letter-spacing: 2px;
-  transition: transform 0.5s ease-in-out;
-}
-
-.input-add-item:focus {
-  box-shadow: 5px 6px 5px #888888;
-  border: 2px solid #e9f3e9;
-  transform: scale(1.03);
-}
