--- src/shared/components/Input/Text/Text.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import './style.scss';
-
-function TextComponent(props) {
-  return (
-    <p className="thing-text">{props.text}</p>
-  )
-}
-export default TextComponent;
