--- src/shared/components/Input/Text/style.scss
@@ -1,6 +1,0 @@
-@import './../../main-styles.scss';
-
-.thing-text {
-  letter-spacing: 2px;
-  color: $thing-text-color;
-}
