+++ src/shared/components/Text/style.scss
@@ -1,0 +1,5 @@
+@import './../../main-styles.scss';
+
+.thing-text {
+  color: $thing-text-color;
+}
