--- src/shared/components/main-styles.scss
@@ -1,15 +1,0 @@
-$main-color: lightgray;
-$bg-color: gainsboro;
-$thing-text-color: gray;
-$title-color: gray;
-$input-text-color: darkgray;
-
-@mixin button-styles {
-  outline: 0;
-  cursor: pointer;
-}
-
-@mixin row-form {
-  display: flex;
-  flex-direction: row;
-}
