+++ src/shared/utils/localStorage.js
@@ -1,0 +1,9 @@
+export function saveItemInLocalStorage(key, value) {
+  const valueStringified = JSON.stringify(value);
+  localStorage.setItem(key, valueStringified);
+}
+
+export function getItemFromLocalStorage(key) {
+  const value = localStorage.getItem(key)
+  return JSON.parse(value);
+}
